@@ -1,26 +1,35 @@
-apply plugin: 'com.android.application'
-apply plugin: 'me.tatarka.retrolambda'
+// apply plugin: 'com.android.application'
+// apply plugin: 'me.tatarka.retrolambda'
+plugins {
+    id 'com.android.application'
+    // id 'me.tatarka.retrolambda'
+}
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 30
+    buildToolsVersion "30.0.3"
+
     aaptOptions.cruncherEnabled = false
     aaptOptions.useNewCruncher = false
+
     defaultConfig {
         applicationId "com.lqr.wechat"
         minSdkVersion 16
-        targetSdkVersion 25
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         multiDexEnabled true
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
