@@ -1,6 +1,6 @@
 package com.lqr.wechat.db.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import org.litepal.crud.DataSupport;
