@@ -1,7 +1,7 @@
 package com.lqr.wechat.widget;
 
 import android.content.Context;
-import android.support.v7.widget.CardView;
+import androidx.cardview.widget.CardView;
 import android.util.AttributeSet;
 
 import com.zhy.autolayout.AutoFrameLayout;
