@@ -2,7 +2,7 @@ package com.lqr.wechat.widget;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.util.TypedValue;
